// src/infra/api/model/inviteUserBody.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Backend Store Check
 * API for storage and management of products and users
 * OpenAPI spec version: 1.0.0
 */
import type { InviteUserBodyRole } from './inviteUserBodyRole';

export type InviteUserBody = {
  email: string;
  branch_id?: string;
  role?: InviteUserBodyRole;
};