// src/infra/api/model/listOwners200DataItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Backend Store Check
 * API for storage and management of products and users
 * OpenAPI spec version: 1.0.0
 */

export type ListOwners200DataItem = {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  business_name: string;
  /** @nullable */
  logo_url: string | null;
  is_active: boolean;
  created_at: string;
  updated_at: string;
};